Account for leap year in 'day of year' validation
The build is failing today (31st December 2020) because 2020 is a
leap year, therefore this year there are 366 days in the year
rather than 365.

Rather than simply increase the allowed range from 365 to 366,
I've added some conditional logic to check different date ranges
depending on if the year is a leap year or not. This should
prevent other bugs from creeping in!